+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Updated table styles to simplify and unify table margins
+- Implemented Search for Students table
 
 ## 0.15.1
 - Updated views for new grade structure